Fetch books by ISBN in a single request
Earlier, two requests were being sent to find a book. The first to fetch
the book's OLID, and the second fetched the details.

The function has been changed to use the newer REST API rather than the
older JSON API, so the book details can be fetched in a single HTTP
call.